- remove preDelete and PostRun hook - extract code related to manifest.status to new package
