+++ tests/integration/controller/manifest/custom_resource_check/suite_test.go
@@ -159,8 +159,7 @@ var _ = BeforeSuite(func() {
 			func(_ context.Context, _ declarativev2.Object) (*declarativev2.ClusterInfo, error) {
 				return &declarativev2.ClusterInfo{Config: authUser.Config()}, nil
 			},
-		), manifest.WithClientCacheKey(), declarativev2.WithPostRun{manifest.PostRunCreateCR},
-		declarativev2.WithPreDelete{manifest.PreDeleteDeleteCR},
+		), manifest.WithClientCacheKey(),
 		declarativev2.WithCustomStateCheck(statecheck.NewManagerStateCheck(statefulChecker, deploymentChecker)))
 
 	err = ctrl.NewControllerManagedBy(mgr).
